Major client code update as well as temp core hack
Client code has been reworked and refactored to make the web services
package more managable as well as have it's own configuration builder.

Core "hack" is a hack for GET request for messages. Previous set up did
not return results that were JSON friendly so hack makes them so. Will
need to double back and think of a possibly cleaner way of handling
this.